Extract Codex source-home control into its own module
Keeps the AgentsPane addition small (+34 net lines) instead of growing an
already-max-lines-suppressed file by ~137 lines.
